move bolt12 request failures to OfferManager
